Added preview readiness check before initializing (#99)
# Pull Request

## 📖 Description

<!--- Provide some background and a description of your work. -->
This pull request checks for the preview to be ready before sending an initialization message. There is a residual issue in the app where the messages are not synced, and the monaco adaptor sends initialization messages that are unnecessary when the monaco adaptor is not doing the updates. This will be addressed in a later code update.

## ✅ Checklist

### General

<!--- Review the list and put an x in the boxes that apply. -->

- [ ] I have added tests for my changes.
- [x] I have tested my changes.
- [ ] I have updated the project documentation to reflect my changes.