Added a location utility for schema  that will resolve URI fragment (#117)
# Pull Request

## 📖 Description

<!--- Provide some background and a description of your work. -->
This simple string manipulation utility will be used by the solution for #111. Intended for use in `design-to-code` utilities and `design-to-code-react` `Form` component.

### 🎫 Issues

<!---
List and link relevant issues here using the keyword "closes"
if this PR will close an issue, eg. closes #411
-->
#111

## ✅ Checklist

### General

<!--- Review the list and put an x in the boxes that apply. -->

- [x] I have added tests for my changes.
- [x] I have tested my changes.
- [ ] I have updated the project documentation to reflect my changes.

## ⏭ Next Steps

<!---
If there is relevant follow-up work to this PR, please list any existing issues or provide brief descriptions of what you would like to do next.
-->
- Allow `$ref` targeting `$defs` to generate example data